add support for mock mode
I was wanting to create a flux mock plugin (and I still do)
but a much easier approach is to just mock the flux handle from
the main flux-burst library here. This will create a faux job
that can be used to test a plugin outside of a context that has
flux. This also means the burst will be isolated (not connecting
to the local cluster) so this should be a mode that is added
to each of the burst plugins for testing.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>